refactor: moved UserThumbnail into components
